+++ miner.go
@@ -19,7 +19,7 @@ type Miner struct {
 	gorm.Model
 	Coin                 string
 	Address              string `gorm:"unique;not null"`
-	Balance              int64
+	Balance              float64
 	LastPaymentTimestamp int64
 }
 
@@ -56,12 +56,12 @@ func (m *Miner) String() string {
 // Payment to store payment attributes
 type Payment struct {
 	Hash      string
-	Value     int64
+	Value     float64
 	Timestamp int64
 }
 
 // NewPayment creates a Payment
-func NewPayment(hash string, value int64, timestamp int64) *Payment {
+func NewPayment(hash string, value float64, timestamp int64) *Payment {
 	return &Payment{
 		Hash:      hash,
 		Value:     value,
